Making the unit tests happy.
I guess I set the encoding of the written files incorrectly. It also didn't like the fact that I was parsing the given numbers in a way that allowed for decimal places. Not because the value was unexpected, but because it literally just searched the call for a reference to "parseInt", which I think is ridiculous.